DataProcessing/requirements.txt
amqp==2.5.2
billiard==3.6.3.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
Django==1.8.2
idna==2.8
importlib-metadata==1.5.0
Jinja2==2.10.3
joblib==0.14.1
kiwisolver==1.1.0
kombu==4.6.7
lxml==3.7.2
MarkupSafe==1.1.1
matplotlib==3.1.2
numpy==1.18.1
pandas==0.25.3
patsy==0.5.1
prettytable==0.7.2
pyecharts==1.6.2
PyMySQL==0.9.3
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.22.0
scikit-learn==0.22.1
scipy==1.4.1
selenium==3.141.0
simplejson==3.17.0
six==1.13.0
sklearn==0.0
SQLAlchemy==1.3.13
statsmodels==0.10.2
urllib3==1.25.7
vine==1.3.0
zipp==3.0.0
